datapath: Move NodeAddressing out from tables
NodeAddressing doesn't make sense to live in the tables package as it
isn't about defining a common StateDB table. It also introduces dependency
on the datapath types package which makes it impossible to use the tables
package types in datapath types.

Eventually whole NodeAddressing will be removed as it is just a wrapper around
LocalNodeStore and Table[Device], and a harmful one at that as it hides the
fact that the underlying data may change over time.
